requirements.txt
fastapi
uvicorn
modal
requests
httpx
python-dotenv
pydantic
certifi
loguru
google-auth
google-auth-oauthlib
google-api-python-client
streamlit
requests
pandas
google-auth
google-api-python-client
